# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Modern, lightweight Kubernetes dashboard API.",
        "title": "K-View API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/oisis/k-view",
            "email": "support@k-view.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/resources/{kind}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generic endpoint to list any Kubernetes resource kind, with support for custom resources and namespace filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "List Resources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource Kind (e.g. pods, deployments, mycustomresources)",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace filter",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "API Group (for CRDs)",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "API Version (for CRDs)",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural (for CRDs)",
                        "name": "plural",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ResourceItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ResourceItem": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}